HomeViewのテストクラスをaccounts/tests.pyからtweets/tests.pyに移動させました。
