+++ tweets/tests.py
@@ -1,9 +1,18 @@
-# from django.test import TestCase
+from django.contrib.auth import get_user_model
+from django.test import TestCase
+from django.urls import reverse
+
+User = get_user_model()
 
 
-# class TestHomeView(TestCase):
-#     def test_success_get(self):
+class TestHomeView(TestCase):
+    def setUp(self):
+        self.user = User.objects.create_user(username="testuser", password="testpassword")
+        self.client.login(username="testuser", password="testpassword")
 
+    def test_success_get(self):
+        response = self.client.get(reverse("tweets:home"))
+        self.assertEqual(response.status_code, 200)
 
 # class TestTweetCreateView(TestCase):
 #     def test_success_get(self):
